+++ src/main/com/ocaml/ide/console/debug/OCamlStupidTreeModel.java
@@ -3,8 +3,10 @@ package com.ocaml.ide.console.debug;
 import com.intellij.ide.structureView.StructureViewTreeElement;
 import com.intellij.ide.util.treeView.smartTree.*;
 import com.intellij.navigation.ItemPresentation;
+import com.ocaml.ide.console.debug.groups.OCamlModuleGroup;
 import com.ocaml.ide.console.debug.groups.TreeElementGroup;
 import com.ocaml.ide.console.debug.groups.TreeElementGroupKind;
+import com.ocaml.ide.console.debug.groups.elements.OCamlTreeElement;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -19,9 +21,9 @@ import java.util.ArrayList;
  */
 public class OCamlStupidTreeModel implements TreeModel {
 
-    public final TreeElementGroup exceptions = new TreeElementGroup(TreeElementGroupKind.EXCEPTION);
-    public final TreeElementGroup types = new TreeElementGroup(TreeElementGroupKind.TYPE);
-    public final TreeElementGroup modules = new TreeElementGroup(TreeElementGroupKind.MODULE);
+    public final TreeElementGroup exceptions = new TreeElementGroup(TreeElementGroupKind.EXCEPTIONS);
+    public final TreeElementGroup types = new TreeElementGroup(TreeElementGroupKind.TYPES);
+    public final TreeElementGroup modules = new OCamlModuleGroup(); // this one is special
     public final TreeElementGroup functions = new TreeElementGroup(TreeElementGroupKind.FUNCTIONS);
     public final TreeElementGroup variables = new TreeElementGroup(TreeElementGroupKind.VARIABLES);
     public TreeElement root = new StructureViewTreeElement() {
@@ -71,8 +73,7 @@ public class OCamlStupidTreeModel implements TreeModel {
     public void remove(Object o) {
         for (TreeElement e : root.getChildren()) {
             if (!(e instanceof TreeElementGroup)) continue;
-            //noinspection SuspiciousMethodCalls
-            ((TreeElementGroup) e).elements.remove(o);
+            ((TreeElementGroup) e).removeElement((OCamlTreeElement) o);
         }
     }