+++ src/main/com/ocaml/ide/console/debug/OCamlVariablesView.java
@@ -12,6 +12,7 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.SimpleToolWindowPanel;
+import com.intellij.openapi.util.Pair;
 import com.intellij.ui.ScrollPaneFactory;
 import com.intellij.ui.tree.AsyncTreeModel;
 import com.intellij.ui.tree.StructureTreeModel;
@@ -20,14 +21,13 @@ import com.ocaml.ide.console.OCamlConsoleView;
 import com.ocaml.ide.console.debug.groups.ShowElementGroupAction;
 import com.ocaml.ide.console.debug.groups.TreeElementGroup;
 import com.ocaml.ide.console.debug.groups.TreeElementGroupKind;
-import com.ocaml.ide.console.debug.groups.elements.OCamlFunctionElement;
 import com.ocaml.ide.console.debug.groups.elements.OCamlTreeElement;
-import com.ocaml.ide.console.debug.groups.elements.OCamlVariableElement;
 import com.ocaml.utils.logs.OCamlLogger;
 import org.jetbrains.annotations.NotNull;
 
 import javax.swing.*;
 import java.awt.*;
+import java.util.List;
 
 /**
  * Panel with the variables
@@ -72,9 +72,9 @@ public class OCamlVariablesView extends SimpleToolWindowPanel implements Disposa
     private @NotNull AnAction createSettingsActionGroup() {
         var group = new DefaultActionGroup("Variable View Settings", true);
         group.getTemplatePresentation().setIcon(AllIcons.Actions.Show);
-        group.add(new ShowElementGroupAction(TreeElementGroupKind.EXCEPTION, this));
-        group.add(new ShowElementGroupAction(TreeElementGroupKind.MODULE, this));
-        group.add(new ShowElementGroupAction(TreeElementGroupKind.TYPE, this));
+        group.add(new ShowElementGroupAction(TreeElementGroupKind.EXCEPTIONS, this));
+        group.add(new ShowElementGroupAction(TreeElementGroupKind.MODULES, this));
+        group.add(new ShowElementGroupAction(TreeElementGroupKind.TYPES, this));
         group.add(new ShowElementGroupAction(TreeElementGroupKind.FUNCTIONS, this));
         group.add(new ShowElementGroupAction(TreeElementGroupKind.VARIABLES, this));
         return group;
@@ -86,41 +86,28 @@ public class OCamlVariablesView extends SimpleToolWindowPanel implements Disposa
      * @param newEntry lines of the new item (ex: val x : int = 5)
      */
     public void rebuild(@NotNull String newEntry) {
-        if (!newEntry.startsWith("val")) return; // not possible yet
-        // handle "val" -> variables or functions
+        List<Pair<OCamlTreeElement, TreeElementGroupKind>> results = OCamlREPLOutputParser.parse(newEntry);
+        if (results == null) return;
 
-        newEntry = newEntry.trim();
-        // todo: log
-        if (newEntry.endsWith("<fun>")) LOG.debug("adding a function:" + newEntry);
-        else LOG.debug("adding a variable:" + newEntry);
-
-        // todo: const + parser
-
-        // find name
-        int val = newEntry.indexOf("val");
-        int colon = newEntry.indexOf(":");
-        int equals = newEntry.indexOf("=");
-        String name = newEntry.substring(val + 3 + 1, colon - 1);
-
-        // find type
-        String type = newEntry.substring(colon + 1, equals - 1);
-
-        OCamlTreeElement element;
-        TreeElementGroup group;
-        if (!newEntry.endsWith("<fun>")) {
-            // find value
-            String value = newEntry.substring(equals + 1);
-            element = new OCamlVariableElement(name, value, type);
-            group = treeModel.variables;
-        } else {
-            element = new OCamlFunctionElement(name, type);
-            group = treeModel.functions;
+        for (Pair<OCamlTreeElement, TreeElementGroupKind> g : results) {
+            // add
+            TreeElementGroup group = null;
+            switch (g.second) {
+                case EXCEPTIONS: group = treeModel.exceptions; break;
+                case MODULES: group = treeModel.modules; break;
+                case TYPES: group = treeModel.types; break;
+                case FUNCTIONS: group = treeModel.functions; break;
+                case VARIABLES: group = treeModel.variables; break;
+            }
+            if (group == null) {
+                LOG.error("Unknown kind:"+g.second);
+                continue;
+            }
+            // remove
+            treeModel.remove(g.first);
+            // add
+            group.addElement(g.first);
         }
-
-        // remove
-        treeModel.remove(element);
-        // add
-        group.elements.add(element);
 
         invalidateTree();
     }