+++ src/main/com/ocaml/ide/console/debug/groups/ShowElementGroupAction.java
@@ -21,9 +21,9 @@ public class ShowElementGroupAction extends DumbAwareToggleAction {
 
         OCamlStupidTreeModel treeModel = view.getTreeModel();
         switch (kind) {
-            case EXCEPTION: target = treeModel.exceptions; break;
-            case MODULE: target = treeModel.modules; break;
-            case TYPE: target = treeModel.types; break;
+            case EXCEPTIONS: target = treeModel.exceptions; break;
+            case MODULES: target = treeModel.modules; break;
+            case TYPES: target = treeModel.types; break;
             case FUNCTIONS: target = treeModel.functions; break;
             case VARIABLES: target = treeModel.variables; break;
             default: throw new IllegalStateException("Invalid kind:'"+kind.displayName+"'.");