+++ src/main/com/ocaml/ide/console/debug/groups/TreeElementGroupKind.java
@@ -2,18 +2,24 @@ package com.ocaml.ide.console.debug.groups;
 
 import com.ocaml.OCamlBundle;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.PropertyKey;
 
+// it's not safe to add a value here
+// ensure that you know were they are used before adding one
+// - removeElement
+// - parser + tests
+// - create an element in elements/
+// - ShowElementGroupAction
+// - OCamlVariablesView#rebuild
 public enum TreeElementGroupKind {
-    EXCEPTION("repl.variable.view.exceptions"),
-    MODULE("repl.variable.view.modules"),
-    TYPE("repl.variable.view.types"),
-    FUNCTIONS("repl.variable.view.functions"),
-    VARIABLES("repl.variable.view.variables");
+    EXCEPTIONS(OCamlBundle.message("repl.variable.view.exceptions")),
+    MODULES(OCamlBundle.message("repl.variable.view.modules")),
+    TYPES(OCamlBundle.message("repl.variable.view.types")),
+    FUNCTIONS(OCamlBundle.message("repl.variable.view.functions")),
+    VARIABLES(OCamlBundle.message("repl.variable.view.variables"));
 
     public final String displayName;
 
-    TreeElementGroupKind(@NotNull @PropertyKey(resourceBundle = OCamlBundle.BUNDLE) String key) {
-        this.displayName = OCamlBundle.message(key);
+    TreeElementGroupKind(@NotNull String name) {
+        this.displayName = name;
     }
 }