+++ test/testData/console/issue57_eof.ml
@@ -1,0 +1,2 @@
+(* fixed, cursor after the last quote -> CTRL-ENTER failed *)
+let x = "Hello, World!@."