+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The plugin is experimental. You should implement any feature directly in the [Re
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7]
+### Added
+- ✨ Can show/hide the variable view, can be docked next to the file (#40)
+- ✨ preview auto-selected statement (CTRL-ENTER) if needed (#56)
+
+### Changed
+- ✨ Fix issues with the REPL console (#37, #57, #60)
+
 ## [0.0.6] - 2022-02-05
 ### Internal
 - Keeping a Changelog